nanorobot-vision/requirements.txt
# requirements.txt
# Place your dependencies here
# requirements.txt
cellpose==2.2.0
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.0.74
scikit-image==0.21.0
numpy==1.24.0
Pillow==9.5.0
tqdm==4.65.0
pandas==2.0.3